Moved distanceCalculator.py to core and updated dependent imports.
